Move getIgnoredUpgrades and upgradePackageDefinitions to /src/lib
